+++ Changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 Version 0.13.0
 --------------
+* render subtype in markdown
 * restructured overarch models
 * added an explicit example of a concept model and view